pyproject.toml
[tool.poetry]
name = "allennlp-shiba-model"
version = "0.1.0"
description = ""
authors = ["Shunsuke KITADA <shunsuke.kitada.0831@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
allennlp = "^2.5.0"
shiba-model = "^0.1.0"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
isort = "^5.9.1"
flake8 = "^3.9.2"
mypy = "^0.910"
pytest = "^6.2.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
